Code for state FIR_MAIN
I've written code for the state FIR_MAIN in n_tap_fir. This state does the main computation that is required of the FIR filter. The slightly altered test bench seems to show the FIR filter working as expected. After doing a convolution operation in Matlab between the input data and coefficients, it seems the first 10 output values are correct, after that there is a slight mismatch in numbers. Still need to find the bug.