Bug found from last commit
From my last commit, 2 bugs were found. Seems like the last value of the multiplication between the coeff_buffer and input_data_buffer was not happening as the script stopped shy of the last length value. The second bug found was with the test script, seems the script was written to send the input data in 10 increments but the value 100 was missing, so that was added.