tested for length of 20
Seems the scripts work for length 10, 15, and 20, thus safe to say it will work for any entered length value. The convolution process seems to work for any length of the input. However, I've noticed the last [entered length of the filter] values of the output are of incorrect value (when checked against Matlab). This is because Matlab pads "0"s to the end of ine input data (it adds entered length of the filter - 1). The same will not be done with this script, maybe in a different check script.