Finalised n_tap_fir + tb
Added a stop_data_load_flag so that the script can transistion to the stop state. Also added 19 padding 0 to the input_data in the test bench to comfirm the logic of the previous stament. The n_tap_fir script works correctly as comfirmed by the conv(x,y) function in Matlab.